refactor: change name from Video to TodaysVideo
