+++ frontend/src/components/TodaysVideo/CollectButton/CollectButton.tsx
